devices as apps
