+++ DjangoControlServer/urls.py
@@ -20,4 +20,7 @@ from django.views.generic import RedirectView
 urlpatterns = [
 	path('admin/', admin.site.urls),
 	path('', include('main.urls')),
+	path('arduino/', include('arduino.urls')),
+	path('mokugo/', include('mokugo.urls')),
+	path('slackbot/', include('slackbot.urls')),
 ]
