+++ apps/arduino/apps.py
@@ -1,0 +1,5 @@
+from django.apps import AppConfig
+
+class ArduinoConfig(AppConfig):
+    default_auto_field = 'django.db.models.BigAutoField'
+    name = 'arduino'
