+++ apps/main/static/main/index.js
@@ -11,6 +11,9 @@ IndexTable.component('index-table', {
 	},
 	props: ['devices'],
 	template: `
+	
+	[[ this.devices ]]
+	
 	<div class="row mb-3">
 		<div class="col"><select v-model="this.addForm['model']" class="form-select">
 			<option value="main.pdmon">PDmon</option>
@@ -88,7 +91,7 @@ IndexTable.component('device-widget', {
 	</div></td>
 	`,
 	mounted () {
-		this.get_device();
+		//this.get_device();
 	},
 	methods : {
 		get_device() {
