+++ apps/mokugo/admin.py
@@ -1,0 +1,4 @@
+from django.contrib import admin
+from .models import mokugo
+
+# Register your models here.
