+++ apps/mokugo/templates/mokugo/mokugo.html
@@ -1,0 +1,11 @@
+{% load static %}
+
+{% block content %}
+<p>{{ moku }}</p>
+
+<!-- div class="container-fluid" id="mokugo">
+	<mokugo v-bind:mokugo="{{ device }}"></mokugo>
+</div>
+
+<script src="{% static 'main/mokugo.js' %}"></script -->
+{% endblock content %}
