+++ apps/mokugo/urls.py
@@ -1,0 +1,6 @@
+from django.urls import path, include
+from . import views
+
+urlpatterns = [
+	path('<str:mokugo_name>/', views.mokugo, name='mokugo'),
+]
