+++ apps/slackbot/static/slackbot/slackbot.js
