+++ apps/slackbot/templates/slackbot/slackbot.html
@@ -1,0 +1,13 @@
+{% extends "main/masterpage.html" %}
+{% load static %}
+
+{% block content %}
+
+<p>{{ slackbot }}</p>
+
+<!-- div class="container-fluid" id="slackbot">
+	<slackbot></slackbot>
+</div>
+
+<script src="{% static 'main/slackbot.js' %}?rand=23423422"></script -->
+{% endblock content %}
