--- main/admin.py
@@ -1,7 +1,0 @@
-from django.contrib import admin
-from .models import PDmon, Tctrl, MokuGo
-
-# Register your models here.
-admin.site.register(PDmon)
-admin.site.register(Tctrl)
-admin.site.register(MokuGo)
