--- main/templates/main/mokugo.html
@@ -1,14 +1,0 @@
-{% extends "main/masterpage.html" %}
-{% load static %}
-
-{% block content %}
-{% csrf_token %}
-
-<p>{{ message }}</p>
-
-<div class="container-fluid" id="mokugo">
-	<mokugo v-bind:mokugo="{{ device }}"></mokugo>
-</div>
-
-<script src="{% static 'main/mokugo.js' %}"></script>
-{% endblock content %}
